Clean up I2S PIO program selection, automatic swap (#3183)
Redo the multiple convoluted ifs + nested trinaries into a simpler
switch() statement to make it easier to work on the library.

Remove all hardcoded "swap" versions of PIO programs because they
are a headache and need to manually be sync'd with the normal pinout
programs.  Instead, dynamically swap the sideset pins from the
single normal-mode version.  Removes 1/2 of the PIO programs.