SDK 2.2.1-develop, enable MIDI, add IRQ slots (#3190)
Fixes #3189

Increase the shared IRQ handler slots by 2 to account for the RP2350
doorbells taking up 2 of them (since we want to allow users to have their
own doorbell handlers).

Move to the latest -develop version of the SDK since we need to rebuild
the SDK anyway.

Enable TinyUSB MIDI support.  Right now there's not a library for it. but
with this folks like @madias123 or others who have asked for it can start
the work.