Fix USB descriptors for multiple interface devices (#3193)
For simple 1-interface devices, there is a fixed position in the USB
interface descriptor where the interface lives and we can hardcode
an updater in the USB handler.  However, for devices like CDC(Serial)
or MIDI there are multiple interfaces and the interface IDs need to
be stuffed inside the actual descriptor.

To allow this, instead of just taking a fixed set of bytes and
blindly patching in USB, make all interfaces do their own memcpy
from a local version, updated with the new dynamic interface id,
via a callback.

Add a simple USBClass callback to handle the normal 1-interface
case.

Fix MSD (SingleFileDrive, FatFSUSB) to only request the correct, single
interface used.