Stability fix for IRQ-based Ethernet
Under heavy load it is possible for a "packet available" interrupt to
happen while another LWIP call is underway.  In that case it is not
safe to inject a packet at IRQ time.

When that happens, now set a flag and disable IRQs and wait until
the LWIP call is done.  When the call finishes, it will re-enable
IRQs and the IRQ callback will be called again, but this time the
inlwip flag will be unset and the ethernet packet inject safe.

I've run over 9000 W5100::begin/WiFiCliet read fortune/W5100::::end
stress (no delays anywhere) cycles without incident.  W/o this patch,
LWIP state got corrupted in a matter of minutes.