Disable IRQs on RP2040 during HW division
The divider is a shared HW resource, and when an IRQ comes in (i.e.
when a packet is processed by LWIP and the user's callbacks) its state can
be corrupted silently and randomly.

Change the Pico-SDK defaults to disable IRQs during division operations,
avoiding the issue by disallowing the LWIP callback to happen until after
division is completed.

Fixes #3212