Override static NCM memory allocation
The NCM buffers are only accessed in a single file and under very
limited circumstances.  Use a SDKOVERRIDE to make a weak ncm_device
with no memory usage, and put the TUSB ncd_device inside the lwip_usb
folder.

For apps which don't use NCM, this saves >6K of RAM.  For apps which use
NCM, there should be no change.