Fix HID endpoint size, remove 512b MSC (#3277)
The HID endpoint size was set incorrectly, update.  No need for new
libraries because the only file that uses the define is in tusb-hid
and will be built dynamically.

The RP2040 endpoints seem to be limited to 64b (also seen in the
tinyusb examples), so don't extend MSC_EP to 512b.